Move abbreviations to a different file
