+++ lib/index.js
@@ -1,30 +1,6 @@
-var abbrev = require('abbrev');
-var mout = require('mout');
 var commands = require('./commands');
 var pkg = require('../package.json');
-
-var abbreviations = abbrev(expandNames(commands));
-abbreviations.i = 'install';
-abbreviations.rm = 'uninstall';
-abbreviations.unlink = 'uninstall';
-abbreviations.ls = 'list';
-
-function expandNames(obj, prefix, stack) {
-    prefix = prefix || '';
-    stack = stack || [];
-
-    mout.object.forOwn(obj, function (value, name) {
-        name = prefix + name;
-
-        stack.push(name);
-
-        if (typeof value === 'object' && !value.line) {
-            expandNames(value, name + ' ', stack);
-        }
-    });
-
-    return stack;
-}
+var abbreviations = require('./abbreviations')(commands);
 
 function clearRuntimeCache() {
     // Note that in edge cases, some architecture components instance's
