+++ test/commands/index.js
@@ -15,7 +15,5 @@ describe('integration tests', function () {
     require('./uninstall');
     require('./update');
     require('./version');
-
-    // run last because it changes defaults
     require('./bower');
 });
